Update names of reterieve and store book classes
