+++ Gemfile
@@ -4,5 +4,5 @@ git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
 
 # gem "rails"
 
+gem 'json', '2.6.1'
 gem 'rubocop', '>= 1.0', '< 2.0'
-gem 'json' , '2.6.1'
