+++ bin/classes/actual_books.rb
@@ -6,5 +6,4 @@ class ActualBook
     @author = author
     @rental = []
   end
-
 end
