+++ bin/methods/added_msg.rb
@@ -1,5 +1,6 @@
 require './utilities/clearing_terminal'
 
 def added_msg(input)
+  clear
   puts "#{input} created successfully ✅", "\n"
 end
